Relocate apps.projects
