+++ bcs-app/backend/apis/projects/views.py
@@ -15,7 +15,7 @@ from rest_framework.response import Response
 
 from backend.apis.authentication import JWTAuthentication
 from backend.apis.permissions import AccessTokenPermission
-from backend.apps.projects.views import NavProjectPermissionViewSet
+from backend.container_service.projects.views import NavProjectPermissionViewSet
 from backend.resources.project import filter_projects
 from backend.utils.renderers import BKAPIRenderer
