+++ bcs-app/backend/container_service/projects/constants.py
