+++ bcs-app/backend/container_service/projects/models.py
@@ -15,7 +15,7 @@ from django.db import models
 from django.utils import timezone
 from django.utils.translation import ugettext_lazy as _
 
-from backend.apps.projects.constants import StaffInfoStatus
+from backend.container_service.projects.constants import StaffInfoStatus
 from backend.utils.models import BaseModel
