+++ bcs-app/backend/container_service/projects/views.py
@@ -22,9 +22,13 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from backend.activity_log import client
-from backend.apps.projects.utils import get_app_by_user_role, get_application_name, update_bcs_service_for_project
 from backend.components import paas_cc
 from backend.components.iam.permissions import ProjectPermission
+from backend.container_service.projects.utils import (
+    get_app_by_user_role,
+    get_application_name,
+    update_bcs_service_for_project,
+)
 from backend.resources import project as Project
 from backend.utils.basic import normalize_datetime
 from backend.utils.cache import region
